# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  company: string;
  projectType: string;
  budget: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}