# lib/api-zod/src/generated/types/createTestimonialBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTestimonialBody {
  name: string;
  company: string;
  role: string;
  quote: string;
  avatarInitials: string;
  avatarColor?: string;
  sortOrder?: number;
}