# lib/api-zod/src/generated/types/portfolioItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioItem {
  id: number;
  title: string;
  category: string;
  description: string;
  techStack: string[];
  resultMetric: string;
  resultLabel: string;
  accentColor: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}