# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  name: string;
  company: string;
  role: string;
  quote: string;
  avatarInitials: string;
  avatarColor: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}